+++ morpheus-graphql-core/changelog.md
@@ -1,3 +1,13 @@
 # Changelog
 
 ## 0.12.0 Unreleased Changes
+
+## New features
+
+- parser supports implemnets interfaces seperated with empty spaces
+
+  ```gql
+  type T implements A , B C & D {
+  ```
+
+- introspection can render interfaces